Further worked on Meta-Table based serialization. Added support for material and MaterialPassInsatance
Former-commit-id: 61b27c7ba9670d586ce962fba691484f48a09ba4

+++ fancy_core/src/Serializer.cpp
@@ -86,7 +86,7 @@ namespace Fancy { namespace IO {
 
       case EBaseDataType::Float:
       {
-        currJsonVal = *static_cast<uint*>(anObject);
+        currJsonVal = *static_cast<float*>(anObject);
       } break;
 
       case EBaseDataType::Char:
